versi 4

+++ admin/data/suratpesananbarang/proses.php
@@ -81,7 +81,17 @@ else {
         if (isset($_POST['save'])) {
 			$id_fasilitas = mysqli_real_escape_string($koneksi, stripslashes(strip_tags(htmlspecialchars(trim($_POST['id_fasilitas'])))));
             if (isset($id_fasilitas)) {
-        			$nm_fasilitas = mysqli_real_escape_string($koneksi, stripslashes(strip_tags(htmlspecialchars(trim($_POST['nm_fasilitas'])))));
+					
+					$allowed_ext	= array('pjpeg', 'jpg', 'jpeg', 'png');
+					$file_name		= $_FILES['file']['name'];
+					$file_ext		= strtolower(end(explode('.', $file_name)));
+					$file_size		= $_FILES['file']['size'];
+					$file_tmp		= $_FILES['file']['tmp_name'];
+					$acak        	= rand(1,99);
+					$nama_gambar 	= $acak.$file_name;
+					
+					$id_jnsfasilitas = mysqli_real_escape_string($koneksi, stripslashes(strip_tags(htmlspecialchars(trim($_POST['id_jnsfasilitas'])))));
+					$nm_fasilitas = mysqli_real_escape_string($koneksi, stripslashes(strip_tags(htmlspecialchars(trim($_POST['nm_fasilitas'])))));
         			$nm_kegiatan = mysqli_real_escape_string($koneksi, stripslashes(strip_tags(htmlspecialchars(trim($_POST['nm_kegiatan'])))));
         			$nm_penyedia = mysqli_real_escape_string($koneksi, stripslashes(strip_tags(htmlspecialchars(trim($_POST['nm_penyedia'])))));
 					$id_peta = mysqli_real_escape_string($koneksi, stripslashes(strip_tags(htmlspecialchars(trim($_POST['id_peta'])))));
@@ -111,18 +121,9 @@ else {
 						total_harga = '$total_harga', 
 						keterangan = '$keterangan' 
 						WHERE id_fasilitas = '$id_fasilitas'");
-
-						if($updt == true){
-										
-						echo "<meta http-equiv='refresh' content='0; url=../../main.php?page=suratpesananbarang&alert=2'>";
-						exit;
-						}else{
-							echo "<meta http-equiv='refresh' content='0; url=../../main.php?page=suratpesananbarang&act=edit&id=$id_fasilitas'>";
-						exit;
-						}
-
-						}else{
-									
+					
+					if(empty($file_tmp)){
+							
 						$updt = mysqli_query($koneksi, "UPDATE tbl_fasilitas SET 
 						nm_fasilitas = '$nm_fasilitas',
 						nm_kegiatan = '$nm_kegiatan', 
@@ -138,9 +139,46 @@ else {
 						total_harga = '$total_harga', 
 						keterangan = '$keterangan' 
 						WHERE id_fasilitas = '$id_fasilitas'");
+						if($updt == true){
+							echo "<meta http-equiv='refresh' content='0; url=../../main.php?page=suratpesananbarang&alert=2'>";
+							exit;
+						}else{
+							echo "<meta http-equiv='refresh' content='0; url=../../main.php?page=suratpesananbarang&act=edit&id=$id_fasilitas'>";
+							exit;
+						}
+									
+					}else{
+									
+						if ($file_ext =='php' || $file_ext =='*.php' || $file_ext =='*.php.jpg' || $file_ext =='*.php.jpeg' || $file_ext =='.php.png' || $file_ext =='.php.pjeg'){
+							echo "<script>window.alert('Upload Gagal, Pastikan File yang di Upload tidak bertipe *.PHP');window.location=('../../main.php?page=fasilitas&alert=5')</script>";
+						}else{
+							if(in_array($file_ext, $allowed_ext) == true){
+								$nm_file = md5($nama_gambar);
+								$fl = $nm_file.'.'.$file_ext;
+								$lokasi = '../../../lib/img/fasilitas/';
+								$folder = "../../../lib/img/fasilitas/";
+								$ukuran = 200;
+								$updt = mysqli_query($koneksi, "UPDATE tbl_fasilitas SET 
+									nm_fasilitas = '$nm_fasilitas',
+									nm_kegiatan = '$nm_kegiatan', 
+									nm_penyedia = '$nm_penyedia',
+									id_peta = '$id_peta', 
+									isi_surat = '$isi_surat', 
+									lo_fasilitas = '$lo_fasilitas', 
+									la_fasilitas = '$la_fasilitas', 
+									nm_barang = '$nm_barang', 
+									stn_barang = '$stn_barang', 
+									jml_barang = '$jml_barang', 
+									hrg_satuan = '$hrg_satuan',
+									total_harga = '$total_harga', 
+									keterangan = '$keterangan' 
+									WHERE id_fasilitas = '$id_fasilitas'");
 
 								if($updt == true){
-
+									
+									UploadFotox($fl, $folder, $ukuran);
+									unlink($folder.$_POST['gb_l']);	
+									unlink($folder.'small_'.$_POST['gb_l']);	
 									echo "<meta http-equiv='refresh' content='0; url=../../main.php?page=suratpesananbarang&alert=2'>";
 									exit;
 								}else{
@@ -158,14 +196,18 @@ else {
         }else { echo "<meta http-equiv='refresh' content='0; url=../../main.php?page=suratpesananbarang&alert=4'>"; exit; }
     }
 
-	 elseif ($_GET['act']=='delete') {
+    elseif ($_GET['act']=='delete') {
 		$id = mysqli_real_escape_string($koneksi, stripslashes(strip_tags(htmlspecialchars(trim($_GET['id'])))));
         if (isset($id)) {
             
-			$agb = mysqli_query($koneksi,"SELECT id_fasilitas FROM tbl_fasilitas WHERE id_fasilitas = '$id'") 
+			$agb = mysqli_query($koneksi,"SELECT gb_fasilitas,id_fasilitas FROM tbl_fasilitas WHERE id_fasilitas = '$id'") 
             or die('Ada kesalahan pada query tampil data : '.mysqli_error());
 			$datax = mysqli_fetch_assoc($agb);
             if($datax['id_fasilitas'] == true){
+				if ($datax['gb_fasilitas'] != ''){
+					$folder = "../../../lib/img/fasilitas";
+					unlink($folder.$datax['gb_fasilitas']);	
+					unlink($folder."small_".$datax['gb_fasilitas']);	
 					
 					$query = mysqli_query($koneksi,"DELETE FROM tbl_fasilitas WHERE id_fasilitas = '$id'")
                             or die('Ada kesalahan pada query delete : '.mysqli_error($koneksi));
@@ -182,5 +224,7 @@ else {
             
         }else{ echo "<meta http-equiv='refresh' content='0; url=../../main.php?page=suratpesananbarang&alert=6'>"; exit; }
     }
+	}
+	
 }       
 ?>
